+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to Rust's notion of
 - Migrated to `zcash_note_encryption 0.2`.
 - `orchard::builder::Builder::{add_spend, add_output}` now use 
   concrete error types instead of `&'static str`s. 
+- `orchard::builder::Error` is now `BuildError` to differentiate from
+  new error types
 
 ## [0.2.0] - 2022-06-24
 ### Added